Fix/double to float

+++ io-type/core/src/commonMain/kotlin/com/wsr/conv/ColumnExt.kt
@@ -2,10 +2,10 @@ package com.wsr.conv
 
 import com.wsr.IOType
 
-fun List<IOType.D2>.toColumn(kernel: Int, stride: Int = 1, padding: Int = 0): Array<DoubleArray> {
+fun List<IOType.D2>.toColumn(kernel: Int, stride: Int = 1, padding: Int = 0): Array<FloatArray> {
     val (channel, inputSize) = first().shape
     val output = (inputSize - kernel + 2 * padding) / stride + 1
-    val result = Array(this.size * output) { DoubleArray(kernel * channel) }
+    val result = Array(this.size * output) { FloatArray(kernel * channel) }
     this.forEachIndexed { index, ioType ->
         for (o in 0 until output) {
             val row = index * output + o
