+++ io-type/core/src/commonMain/kotlin/com/wsr/power/SqrtExt.kt
@@ -2,7 +2,7 @@ package com.wsr.power
 
 import com.wsr.IOType
 
-fun IOType.D1.sqrt(e: Double = 1e-7): IOType.D1 {
+fun IOType.D1.sqrt(e: Float = 1e-7f): IOType.D1 {
     val result = this.value.copyOf()
     for (i in result.indices) {
         result[i] = kotlin.math.sqrt(result[i] + e)
@@ -10,7 +10,7 @@ fun IOType.D1.sqrt(e: Double = 1e-7): IOType.D1 {
     return IOType.d1(result)
 }
 
-fun IOType.D2.sqrt(e: Double = 1e-7): IOType.D2 {
+fun IOType.D2.sqrt(e: Float = 1e-7f): IOType.D2 {
     val result = this.value.copyOf()
     for (i in result.indices) {
         result[i] = kotlin.math.sqrt(result[i] + e)
@@ -18,7 +18,7 @@ fun IOType.D2.sqrt(e: Double = 1e-7): IOType.D2 {
     return IOType.d2(shape, result)
 }
 
-fun IOType.D3.sqrt(e: Double = 1e-7): IOType.D3 {
+fun IOType.D3.sqrt(e: Float = 1e-7f): IOType.D3 {
     val result = this.value.copyOf()
     for (i in result.indices) {
         result[i] = kotlin.math.sqrt(result[i] + e)
